4GB-1 messages may require >4GB rdma windows due to not being page-aligned. Pass region length as 64bit on register to avoid truncation.
4GB-1 messages work fine now :)



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/open-mx/trunk@1481 f1ba3bf5-cb5c-402b-92a9-7c6bdc83a356

+++ libopen-mx/omx_large.c
@@ -257,12 +257,12 @@ omx__get_contigous_region(struct omx_endpoint *ep,
   struct omx_cmd_region_segment *rsegs;
   omx_return_t ret;
   uint64_t vaddr;
-  uint32_t rdma_length;
+  uint64_t rdma_length;
   uint16_t offset;
 
   vaddr = ((uintptr_t) buffer) & ~4095;
   offset = ((uintptr_t) buffer) & 4095;
-  rdma_length = (offset + length + 4095) & ~4095;
+  rdma_length = ((uint64_t) offset + (uint64_t) length + 4095) & ~4095;
 
   if (reserver)
     omx__debug_printf(LARGE, "need a region reserved for object %p\n", reserver);
